+++ packages/dns/src/index.ts
@@ -13,11 +13,11 @@ export type {
 } from './profiles/DnsProfile'
 export { DEFAULT_FAKE_IP_FILTER } from './profiles/FakeIpFilter'
 export { DnsProfilePresets } from './profiles/DnsProfilePresets'
-export { buildAndroidDnsProfile } from './profiles/AndroidDnsProfile'
-export type { AndroidDnsProfileOptions } from './profiles/AndroidDnsProfile'
+export { applyAndroidDnsPolicy, buildAndroidDnsProfile } from './profiles/AndroidDnsProfile'
+export type { AndroidDnsPolicyOptions, AndroidDnsProfileOptions } from './profiles/AndroidDnsProfile'
 
 export type { DnsCompiler, CompiledDnsOutput, DnsCompilationMetadata } from './compiler/DnsCompiler'
-export { MihomoDnsCompiler } from './compiler/MihomoDnsCompiler'
+export { MihomoDnsCompiler, resolveDnsIpv6Enabled } from './compiler/MihomoDnsCompiler'
 
 export { DnsValidator } from './manager/DnsValidator'
 export type { DnsValidationResult, DnsValidationError } from './manager/DnsValidator'